FDG-9788 reset
