+++ src/components/data-preview/data-preview-filter-section/data-preview-column-filter/data-preview-column-filter.module.scss
@@ -22,5 +22,5 @@ $filter-width: $dropdown-width - $search-width;
 }
 
 .noMatch {
-  margin: 0 1rem;
+  margin: 1rem 1rem;
 }
